Move event, event_at, event_for, event_exact to builder/event module
